Backport E2E test refactoring
